Skylog and AtlasLoader

+++ Teuria/Core/UserInterface/RustyFileDialog.cs
@@ -1,13 +1,11 @@
-using System;
-using System.IO;
 using System.Runtime.InteropServices;
 using Teuria.Unsafe;
 
 namespace Teuria;
 
-public class RustyFileDialog 
+public class RustyFileDialog
 {
-    public static string Open(string directory = "/") 
+    public static string Open(string directory = "/")
     {
         var strPath = NativeFunctions.open_file(directory);
         if (strPath.error == 64)
@@ -16,19 +14,19 @@ public class RustyFileDialog
         return null;
     }
 
-    public static string Open(string directory = "/", string[] filters = null) 
+    public static string Open(string directory = "/", string[] filters = null)
     {
-        if (filters == null) 
+        if (filters == null)
             return Open(directory);
-        
-        var strPath = NativeFunctions.open_file_with_filter(directory, filters, ((uint)filters.Length));
+
+        var strPath = NativeFunctions.open_file_with_filter(directory, filters, ((nuint)filters.Length));
         if (strPath.error == 64)
             return Marshal.PtrToStringAnsi(strPath.path);
 
         return null;
     }
 
-    public static string Save(string directory = "/") 
+    public static string Save(string directory = "/")
     {
         var strPath = NativeFunctions.save_file(directory);
         if (strPath.error == 64)
@@ -36,12 +34,12 @@ public class RustyFileDialog
         return null;
     }
 
-    public static string Save(string directory = "/", string[] filters = null) 
+    public static string Save(string directory = "/", string[] filters = null)
     {
-        if (filters == null)  
+        if (filters == null)
             return Save(directory);
-        
-        var strPath = NativeFunctions.save_file_with_filter(directory, filters, ((uint)filters.Length));
+
+        var strPath = NativeFunctions.save_file_with_filter(directory, filters, ((nuint)filters.Length));
         if (strPath.error == 64)
             return Marshal.PtrToStringAnsi(strPath.path);
         return null;
